requirements.txt
fastapi
uvicorn
pydantic
pyttsx3
requests
streamlit
